refactor: change structure of the main-page
