+++ src/components/Main-page/Footer/RSLogoSvg.tsx
