+++ src/components/Main-page/Header/AutorizationSvg.tsx
