+++ src/components/Main-page/Mock/Auth.tsx
