+++ src/components/Main-page/Mock/Dictionary.tsx
