+++ src/components/Main-page/Mock/Statistic.tsx
